+++ services/report/prisma/migrations/20241114071514_1_5_0_beta_1/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "TaskPreset" ADD COLUMN     "hidden" BOOLEAN NOT NULL DEFAULT false;
+
+-- AlterTable
+ALTER TABLE "Template" ADD COLUMN     "hidden" BOOLEAN NOT NULL DEFAULT false;